# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/gear/helmets": {
      "get": {
        "description": "Retorna cascos activos con sus variantes, tallas y precio mínimo por tienda. Soporta múltiples filtros.",
        "operationId": "HelmetModelsController_findAll",
        "parameters": [
          {
            "name": "brandSlug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "full_face",
                  "modular",
                  "open_face",
                  "half",
                  "off_road"
                ]
              }
            }
          },
          {
            "name": "shellMaterial",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "polycarbonate",
                  "fiberglass",
                  "carbon_fiber",
                  "aramid_fiber"
                ]
              }
            }
          },
          {
            "name": "closureType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "double_d_ring",
                "micrometric",
                "quick_release",
                "magnetic",
                "pinch_buckle"
              ]
            }
          },
          {
            "name": "visorPinlockCompatible",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "pinlock_30",
                  "pinlock_70",
                  "pinlock_70_max_vision",
                  "pinlock_120",
                  "pinlock_120_max_vision",
                  "pinlock_120_max_vision_xlt",
                  "pinlock_evo",
                  "pinlock_protectint",
                  "pinlock_max_vision",
                  "pinlock_ready",
                  "not_compatible"
                ]
              }
            }
          },
          {
            "name": "visorPinlockIncluded",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "tearOffCompatible",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "certification",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "dot",
                  "ece_2206",
                  "ece_2205",
                  "snell_m2020",
                  "snell_m2025",
                  "fim_frhphe_01",
                  "fim_frhphe_02"
                ]
              }
            }
          },
          {
            "name": "minShellSizes",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "finish",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "gloss",
                  "matte",
                  "satin"
                ]
              }
            }
          },
          {
            "name": "colorFamily",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "black",
                  "white",
                  "red",
                  "blue",
                  "green",
                  "yellow",
                  "orange",
                  "pink",
                  "purple",
                  "grey",
                  "silver",
                  "gold",
                  "brown",
                  "carbon",
                  "multicolor",
                  "bronze",
                  "hi_vis",
                  "iridescent"
                ]
              }
            }
          },
          {
            "name": "mono",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sunVisor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "intercomReady",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minSafetyRating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "maxWeightGrams",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "emergencyRelease",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de cascos"
          }
        },
        "summary": "Listar cascos",
        "tags": [
          "Helmets"
        ]
      }
    },
    "/gear/helmets/by-slug/{brandSlug}/{modelSlug}": {
      "get": {
        "description": "Retorna el detalle completo de un casco con todas sus variantes y tallas.",
        "operationId": "HelmetModelsController_findOne",
        "parameters": [
          {
            "name": "brandSlug",
            "required": true,
            "in": "path",
            "description": "Slug de la marca",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modelSlug",
            "required": true,
            "in": "path",
            "description": "Slug del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del casco"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "summary": "Obtener casco por slug",
        "tags": [
          "Helmets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moto Gear Picker API",
    "description": "API para selección y gestión de equipamiento de moto",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}